Started Work on Hipermercado
Insertion happens at 2 stages, insertion of a new valid client/product,
as read from the catalog files, and then insertion of a sale, which is
only valid if the product/client has been inserted previously. Works
Nicely

PS: MetaSale.java total$ is not being calculated as price * amount,
should it?